Move sessions collection to external.database.
