+++ components/server/src/database/reports.py
@@ -5,11 +5,12 @@ from typing import Any, cast
 import pymongo
 from pymongo.database import Database
 
+from external.database import sessions
 from model.report import Report
 from server_utilities.functions import iso_timestamp, unique
 from server_utilities.type import Change, MetricId, ReportId, SubjectId
+
 from .filters import DOES_EXIST, DOES_NOT_EXIST
-from . import sessions
 
 
 # Sort order:
