+++ components/server/src/external/routes/auth.py
@@ -15,10 +15,11 @@ from ldap3 import ALL, Connection, Server
 from ldap3.core import exceptions
 from pymongo.database import Database
 
-from database import sessions
 from initialization.secrets import EXPORT_FIELDS_KEYS_NAME
 from server_utilities.functions import uuid
 from server_utilities.type import SessionId
+
+from ..database import sessions
 
 
 def create_session(database: Database, username: str, email: str) -> datetime:
