+++ components/server/src/external/routes/plugins/auth_plugin.py
@@ -4,9 +4,11 @@ import logging
 
 import bottle
 
-from database import sessions
 from database.reports import latest_reports_overview
 from model.session import Session
+
+from ...database import sessions
+
 
 EDIT_REPORT_PERMISSION = "edit_reports"
 EDIT_ENTITY_PERMISSION = "edit_entities"
